Move infowindow inside the map for loop
Move infowindow inside the map for loop. In order for all infos
locations to appear, define the position and info on the map first then
create the marker (holding the postions) and add infowindow to the
marker. However now all infowindow
appear in the same last location in the loop